--- src/test/resources/com/cloudbees/jenkins/plugins/sshagent/sshkey
@@ -1,51 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIJKAIBAAKCAgEAxb2ZLz9lDR/aBeCY0P+cNgJoRJ/QF72HWUxqqAHFIQNHBwqu
-JR1Yuh/uhWD2t3D2vVP216CGZaH6fm7GCaMJddS8JrzgYuDwyY5t4bxannJEDtAU
-kBYd3UjjssBliOLh7+2ZeVZnA84QnqD0Vo4v8x2G4rf0dNcVRUEnSO5DxH2XVvfT
-nln2vIYNmODX6H6PagIyLxJwrtYCXHbprBc9CO+3pz9rq1+f5Tp4k4Apes24ZHGe
-FMWJ9OCzISbwtTtSqnUov7qPgfdXEmY54gLXgYz4bto2QFTgrZBiiLmovMmAVv6S
-95Favkk9wchX1I58KTWYroY4PcmtZsGF6JIkqqUgBj4q/AIYr4NQwmsr/JSGLkha
-MXl2UWWD9BrunN3UD9pQRzG2/ocIX98SSGStVCbyj1HVCZR4BvkQsmyJ09BHvw8K
-mUBzcai0F1MTGbl5f283PkpHFm/mCGGXRPwbMMSgjoCJICYc+0CjSut7+0wpsJly
-2WNy+RV4heEKpFjEft4hFiwPxgXW52Del5ccssyFdD45km8Anqv7Yu1k2okpMfpx
-7nWLhZIg7Gol2Wh6wkv+jUvfRGzv10ejuTO6hBGBo2Ic7NBjWtIBYgN1Ml7Gbo1l
-DL7MDJWUVTSk/cBC26So6bWg+lSR+DZGGjL4HFZLbBFDXyEgEXsvSGWAnYMCAwEA
-AQKCAgABB2SFp2nKMLE8uJdy9vGEIm0SdlZp2TWcFFa1uxSggFjdQ4P2c3mjtkjS
-bte5yp+uCoLNSmuImdgo0kC7ihqhf1LfTLpKKWt3UQNnWkiKbIPII/MHcOYE4b01
-j+S3Wp7cVbsns3hjv628dY/v51t+g1fRVC20ce4NZzJiIcM+gXzJGg8C0usCvhvS
-ANF+xzHN1mcWcZ0ee8SVYJUErTZHFqA5AY543M2DLtOzyMXueqWL+D9fUkvOd2DX
-OlGkT8O1WVi17q2Seov8b4W7lFI0vVwJj58apJZqsMaieXzURqQhT7/KA/qZ5N2q
-ZXV3iKpCuKGrvs0KHUOGtlkRBbmlG02nV6NmLH1A3Oi7CCkOO4Y6Gk4kLANHTqs3
-JE1Vsx3PuY4aDx/jIrx6faFa37XBrRABi+G9tNpSjOIG+zv16MD7aymnrn+pqeMj
-8wtd56g2ggLT211vgY4Nv1Sn0uIPKeLqR9tkePWIRhnxqdao32EbZg8z3mE0X4DN
-9psNcdfPFpWqj1IvqWX+SvDhX71hSyh5mt7uZwR6ijxWVuUzTmiPGuo6gWz/LnY2
-c+ZOS6Hnq21zB7oLFlZf+ToTZPS8nEqXsf+tXy3PIRvwRFyUEPrsu8RPP1tjS3Hx
-6be4bD83LmXOPl3XCJKtMuzVdbZhjo4rF0QacTyAKwkMzHqkIQKCAQEA+8jcQ5oB
-Eni/vrDdvTm9ctp8FM/qXb56Zn+b0JITPZU7+7BoH9a2tR9FF4jdITzJwzoBd/7F
-2aB1KvfQL3vSR0oZ+fJmSbRH9C5ZUOAG5FotjrJwn6KmL/NJObDysTG3FoR8Ehsv
-jM0SXviEEM7iioehY72B5vwrDpQDTte43O6N1i8sy6drjTkm1q00r2ZTtEN8lyrl
-Fcl1EnS7m2ARMXQR5J8rXxh9S0Wjkz/O30nkVi1RbwHANlYoZ2G2agX66Xc8xGXi
-vMTCwaKwKRnCzx36mO1xjof9DB3duy5klXMfbtGbdYcq5nnYaqf5F3IDxnGFUgOq
-ffEqktxUhxMekwKCAQEAyQ0bfuOzsz7T3nRlIv620D8eqHinFh/zkERMpBd4BTXw
-H1FVhAcfkCK9m1v/G+2fi7+aspVCiIVv3i4Bkq6OIvkC0Lj5CsILQ08PsyKYce/m
-WqLujeUZLAJJAI8e1qLfEeXkzsTHC2uDDowsXc/Y6jAdWUrKqXP8yrNcufnTiejC
-8NtH+n+jmqh+hk8t42SEOjbiX1XKqmJJWE3qtjYu/DgeOlvUnTRD/b1Rw4+PdkWp
-KMchdHVN3+tHkuU0T/zue/hkq7MB7TJByoAyAMK3NlPVyCRDmmSajEknZjgulnAS
-FZHBu+i2WIpoKi97iUATWsFki01wENVC2GUYQfHrUQKCAQAeWNBLiZVVf55EwpaH
-Rx1RaYoJDCJIWDvyfLDwC5ssEf+fz1xNchilbfZq5sEYT3ZWBHjG2LtKBq/4lNPP
-1VG76++Dk7bGbWc6xBdZyxapkUDqcRFn/DehkeYigPRPbRmPKjq2B/bMNq9fKP7p
-87WUUtKd7eppR718zVQ0vYV3g30u8pKonS42YEwRFeFCvseZ84Rhx/bCkHx+0/kE
-uih4WHALqS+WD6KSXvtfNRKfEQ3nYSDuizx6kBnglANjdgMO/qMhXNJeRCzhH9Wx
-rnwvcIKSyhOkxTk3q6Xc+LOA29ZPCg59tOPz0X3HJSW9sK1Jzsbt4qk68sVZ3BqI
-CULnAoIBAC+sP/zSmWXDXBG/n8uwgkDW9roUAeqpIOPwjgvK1NbHd2y1qIVp+v3+
-ft877T3+FnrYsCHwappcqa4RUqYpZeK9EQHVGpiGco0/TgW4ClXaANE1GUPkMxfQ
-YucTCMj0Tdi8noi4+Emo8KaeblbAVjsRtaParDuKGUMnTLF8n8BvxOm/1U18vpqJ
-oZbPZ54n04LKSNs6SJgmPiYGM8WJC5Ma63ADHmM4ZXbhUWYBwI7W7cmRjklDF3gl
-Hg+i8ciVvScAOz7Lb4WDCT6vLywm/7KOJn8SukT93fet0P3YDJP3dxJOQZYBFFvi
-VdjK8YwV+YDnoqO1dQFI0FTng9z/NEECggEBAL0qqQr775tUfExvMSW+0/5fqcbS
-qsyZAq3DSwkcMt1l0x5WMvv34XsIkrLNhdULMD+u8F00AxxoiF9eb/vSK0U6eofk
-mp5X/V8Bkst5t+2AXYjNma5qMSATTz254flutm7jEhUMpKsuPkkbs/BB3PfB8vmv
-JRXEuWfdBo+CkS1y9Eza6xt70qj7vK1crRiWl2qACdqpx9sBcC2exuCLfkIouCF8
-MHRakEOS3RrZoamyGFwglkk8f7AIrTmELB+JQT3yTcFq8qVijObeS7pOYczOmOxL
-1jePtBMbhyau1wCJw/YFyr9dGG/Wgr9DF4qYTfuI7VQY63SAwcb9OAxzzB0=
------END RSA PRIVATE KEY-----
