--- src/test/resources/com/cloudbees/jenkins/plugins/sshagent/sshkey.pub
@@ -1,1 +1,0 @@
-ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQDFvZkvP2UNH9oF4JjQ/5w2AmhEn9AXvYdZTGqoAcUhA0cHCq4lHVi6H+6FYPa3cPa9U/bXoIZlofp+bsYJowl11LwmvOBi4PDJjm3hvFqeckQO0BSQFh3dSOOywGWI4uHv7Zl5VmcDzhCeoPRWji/zHYbit/R01xVFQSdI7kPEfZdW99OeWfa8hg2Y4Nfofo9qAjIvEnCu1gJcdumsFz0I77enP2urX5/lOniTgCl6zbhkcZ4UxYn04LMhJvC1O1KqdSi/uo+B91cSZjniAteBjPhu2jZAVOCtkGKIuai8yYBW/pL3kVq+ST3ByFfUjnwpNZiuhjg9ya1mwYXokiSqpSAGPir8Ahivg1DCayv8lIYuSFoxeXZRZYP0Gu6c3dQP2lBHMbb+hwhf3xJIZK1UJvKPUdUJlHgG+RCybInT0Ee/DwqZQHNxqLQXUxMZuXl/bzc+SkcWb+YIYZdE/BswxKCOgIkgJhz7QKNK63v7TCmwmXLZY3L5FXiF4QqkWMR+3iEWLA/GBdbnYN6XlxyyzIV0PjmSbwCeq/ti7WTaiSkx+nHudYuFkiDsaiXZaHrCS/6NS99EbO/XR6O5M7qEEYGjYhzs0GNa0gFiA3UyXsZujWUMvswMlZRVNKT9wELbpKjptaD6VJH4NkYaMvgcVktsEUNfISARey9IZYCdgw== noreply@hostname.com
